Rename ContinousLegend, change stories
