+++ packages/react-components-lab/src/components/ContinuousLegend/ContinuousLegend.stories.tsx
@@ -1,8 +1,8 @@
 // also exported from '@storybook/react' if you can deal with breaking changes in 6.1
 import { Meta, Story } from '@storybook/react/types-6-0'; // eslint-disable-line import/no-extraneous-dependencies
-import { Typography } from '@material-ui/core';
+import { Typography, Box } from '@material-ui/core';
 import React from 'react';
-import Legend from './Legend';
+import Legend from './ContinuousLegend';
 import { LegendProps } from './types';
 
 const colors = [
@@ -31,27 +31,44 @@ const colors = [
 export default {
   title: 'Example/Legend',
   component: Legend,
-  argTypes: {
-    backgroundColor: { control: 'color' },
-  },
 } as Meta;
 
 const Template: Story<LegendProps> = (args) => (
-  <div>
+  <>
     <Typography>
       Adds dynamic legend items, based on a color array, and a number range to
       LegendBase
     </Typography>
-    <Legend {...args} />;
-  </div>
+    <Box
+      position="relative"
+      height={500}
+      width={800}
+      p={2}
+      m={2}
+      style={{ backgroundColor: '#cccccc' }}
+    >
+      <Typography>Map</Typography>
+      <Legend {...args} />
+    </Box>
+  </>
 );
 
-export const Primary = Template.bind({});
-Primary.args = {
+export const Minimal = Template.bind({});
+Minimal.args = {
+  colors,
+  min: 0,
+  max: 1,
+};
+
+export const AllProps = Template.bind({});
+AllProps.args = {
   title: 'Velocity',
   unit: 'm/s',
   min: 0,
   max: 4,
   colors,
   maxItems: 10,
+  position: 'topLeft',
+  collapsable: true,
+  defaultCollapsed: true,
 };
