+++ packages/react-components-lab/src/components/ContinuousLegend/ContinuousLegend.tsx
