+++ packages/react-components-lab/src/components/ContinuousLegend/types.ts
